backend/requirements.txt
boto3==1.40.63
alembic==1.16.5
bcrypt==4.0.1
email-validator==2.3.0
fastapi==0.116.1
openai==2.6.1
passlib[bcrypt]==1.7.4
pillow==11.3.0
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.33.1
SQLAlchemy==2.0.43
uvicorn[standard]==0.35.0
